Tried refactoring ship to be inactive instead of having to recreate a ship everytime health reaches 0
As I mentioned, it might be better to just set the ship to inactive,
rather than having to recreate it every time as it might lead to other
issues. Variables are also reset when ship becomes active and there
seems to be no bugs after some testing so let me know if anything goes
wrong.