Omega Theme Song - Preliminary
Added the "music" folder and uploaded the preliminary version of the Omega theme song. At the moment it is only the melody, but I'm very interested in adding some kind of percussion to it and a secondary counter-point melody to it. The WAV file is 4.4MB, which I think is too big. I'll look how to make a smaller file next time.